.agent/skills/snp-simulator/requirements.txt
streamlit
google-genai
pandas
stmol
py3Dmol
python-dotenv
